seconds -> ticks

+++ plugin/src/main/java/com/denizenscript/denizen/objects/properties/entity/EntityConversionTime.java
@@ -90,7 +90,7 @@ public class EntityConversionTime implements Property {
     }
 
     public DurationTag getConversionTime() {
-        return new DurationTag(getEntity().getConversionTime());
+        return new DurationTag((long) getEntity().getConversionTime());
     }
 
     @Override
